requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Migrate==4.1.0
Werkzeug==2.3.7
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
python-dotenv==1.1.0
WTForms==3.2.1
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
blinker==1.6.3
alembic==1.16.2
greenlet==3.0.0
Mako==1.3.10
typing_extensions==4.14.0
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
reportlab==4.4.2
openpyxl==3.1.5
python-docx==1.2.0
Pillow==11.2.1
numpy==2.3.0
pandas==2.3.0
gunicorn==23.0.0
waitress==3.0.2 